+++ src/pages/applicationArea/addApplicationArea/index.js
