+++ src/pages/applicationArea/applicationAreaForm/index.ts
@@ -1,0 +1,2 @@
+import ApplicationAreaForm from '../applicationAreaForm/ApplicationAreaForm';
+export default ApplicationAreaForm;
