+++ src/pages/applicationArea/editApplicationArea/index.js
